refactor: make cbsurge.util functions separate files
