+++ cbsurge/cli.py
@@ -1,6 +1,6 @@
 import logging
 
-from cbsurge.util import setup_logger
+from cbsurge.util.setup_logger import setup_logger
 from cbsurge.admin import admin
 from cbsurge.project import create, list, upload, download, publish
 from cbsurge.initialize import init
