+++ cbsurge/util/chunker.py
@@ -1,0 +1,12 @@
+import itertools
+import logging
+
+logger = logging.getLogger(__name__)
+
+def chunker(iterable, size):
+    it = iter(iterable)
+    while True:
+        chunk = tuple(itertools.islice(it, size))
+        if not chunk:
+            break
+        yield chunk
