+++ cbsurge/util/proj_are_equal.py
@@ -1,0 +1,30 @@
+import logging
+
+from osgeo import osr
+
+logger = logging.getLogger(__name__)
+
+def proj_are_equal(src_srs: osr.SpatialReference = None, dst_srs: osr.SpatialReference = None):
+    """
+    Decides if two projections are equal
+    @param src_srs:  the source projection
+    @param dst_srs: the dst projection
+    @return: bool, True if the source  is different then dst else false
+    If the src is ESPG:4326 or EPSG:3857  returns  False
+    """
+    auth_code_func_name = ".".join(
+        [osr.SpatialReference.GetAuthorityCode.__module__, osr.SpatialReference.GetAuthorityCode.__name__])
+    is_same_func_name = ".".join([osr.SpatialReference.IsSame.__module__, osr.SpatialReference.IsSame.__name__])
+    try:
+        proj_are_equal = int(src_srs.GetAuthorityCode(None)) == int(dst_srs.GetAuthorityCode(None))
+    except Exception as evpe:
+        logger.error(
+            f'Failed to compare src and dst projections using {auth_code_func_name}. Trying using {is_same_func_name}')
+        try:
+            proj_are_equal = bool(src_srs.IsSame(dst_srs))
+        except Exception as evpe1:
+            logger.error(
+                f'Failed to compare src and dst projections using {is_same_func_name}. Error is \n {evpe1}')
+            raise evpe1
+
+    return proj_are_equal
